refactor(oracle): migrate wifi radiotap to wifi plugins
