+++ tools/oracle/tests/test_protocol_plugin_coverage.py
@@ -44,7 +44,6 @@ KNOWN_UNMIGRATED: frozenset[str] = frozenset(
         "ble_radio",
         "dot11",
         "eapol",
-        "radiotap",
         "rsn",
     }
 )
